Power BI_Task1
Hyyy!
Here is a dashboard of global store sales, which include various visualizations like column chart, bar chart, pie chart, tree maps, and like that. Here we have plotted sales with respect to region, country, state, region, markets, and et cetera, which help us easily examine the sales over that region that respect geographical areas. Also, I included a map, which shows us the sales over various geographical areas. Then I added some dropdowns, which help us to filter .not only that, but also added a slicer, which we can use to filter the data for a particular period of time. So that's all about this. Thank you.